pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-expose"
version = "0.1.0"
description = "CLI tool to expose LLMs through messaging clients (Telegram, Discord, Slack, etc.)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "litellm>=1.63.3",
    "python-telegram-bot>=21.10",
    "typer>=0.15.2",
    "rich>=13.9.4",
    "pydantic>=2.10.6",
    "pyyaml>=6.0.2",
]

[project.scripts]
llm-expose = "llm_expose.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["llm_expose*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
